Break out block listener to a start / stop daemon instead of running on startup
