+++ entrypoint.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+python loop.py
