+++ k8s/app.yaml
@@ -17,7 +17,7 @@ spec:
       containers:
       - name: mev-inspect
         image: mev-inspect-py
-        command: [ "/app/run.sh" ]
+        command: [ "/app/entrypoint.sh" ]
         env:
         - name: POSTGRES_USER
           valueFrom:
